Feature/auth into devlopment

+++ src/routes/+page.server.ts
@@ -1,10 +1,16 @@
 import client from '$lib/server';
 import type { PageServerLoad, Actions } from './$types';
 import { fail } from '@sveltejs/kit';
+import { redirect } from '@sveltejs/kit';
 
 //gets all tasks
-export const load: PageServerLoad = async ({ params }) => {
-	const result = await client.GET('/tasks/');
+export const load: PageServerLoad = async ({ locals }) => {
+	if (!locals.user) {
+		redirect(302, '/login');
+	}
+	const result = await client.GET('/tasks/', {
+		params: { header: { userID: locals.user.id } }
+	});
 	let todos: typeof result.data = [];
 	let success = false;
 
@@ -19,9 +25,15 @@ export const load: PageServerLoad = async ({ params }) => {
 };
 
 export const actions = {
-	//deletes all tasks
-	delete: async () => {
-		const result = await client.DELETE('/tasks/');
+	//deletes all tasks for the user
+	delete: async ({ locals }) => {
+		if (!locals.user) {
+			redirect(302, '/login');
+		}
+		const user_id = locals.user.id;
+		const result = await client.DELETE('/tasks/', {
+			params: { header: { userID: locals.user.id } }
+		});
 		let todos: typeof result.data = [];
 		let success = false;
 
@@ -31,73 +43,75 @@ export const actions = {
 		}
 	},
 
-	// add new task
-	create: async (event) => {
-		const formData = await event.request.formData();
+	// add new task for the user
+	create: async ({ locals, request }) => {
+		const formData = await request.formData();
 		const title = formData.get('title')?.toString();
 		const description = formData.get('description')?.toString();
+		if (!locals.user) {
+			redirect(302, '/login');
+		}
 
 		if (!title || !description) {
 			return fail(400, { title, description: description, missing: true });
 		}
 
 		const result = await client.POST('/tasks/', {
+			params: {
+				header: {
+					userID: locals.user.id
+				}
+			},
 			body: {
 				title,
 				description,
 				completed: false,
-				urgent: false,
-				userID: 'Tom'
+				urgent: false
 			}
 		});
 	},
 
 	//delete a specific task by its id
-	deleteTask: async (event) => {
-		const formData = await event.request.formData();
+	deleteTask: async ({ locals, request }) => {
+		const formData = await request.formData();
 		const id = formData.get('_id')?.toString();
+		if (!locals.user) {
+			redirect(302, '/login');
+		}
 		if (!id) {
 			return fail(400, { id, missing: true });
 		}
 		const result = await client.DELETE('/tasks/{id}', {
-			params: { path: { id } }
+			params: {
+				path: { id },
+				header: { userID: locals.user.id }
+			}
 		});
-	},
-
-	//update a specific task by its id
-	updateCompleted: async ({ request }) => {
-		const formData = await request.formData();
-		const id = formData.get('_id')?.toString();
-		const completed = formData.get('completed') === 'on' ? true : false;
-
-		if (completed === null || !id) {
-			return fail(400, { id, completed, missing: true });
-		}
-		const result = await client.PATCH(`/tasks/{id}`, {
-			body: {
-				completed
-			},
-			params: { path: { id } }
-		});
-
-		return { success: true };
 	},
 
 	//update a specific task title or description by its id
-	updateTask: async ({ request }) => {
+	updateTask: async ({ request, locals }) => {
 		const formData = await request.formData();
 		const id = formData.get('_id')?.toString();
 		const title = formData.get('title')?.toString();
 		const description = formData.get('description')?.toString();
+		const completed = formData.get('completed') === 'on' ? true : false;
+		if (!locals.user) {
+			redirect(302, '/login');
+		}
 		if (title === null || description === null || !id) {
 			return fail(400, { id, title, description, missing: true });
 		}
 		const result = await client.PATCH(`/tasks/{id}`, {
 			body: {
 				title,
-				description
+				description,
+				completed
 			},
-			params: { path: { id } }
+			params: {
+				path: { id },
+				header: { userID: locals.user.id }
+			}
 		});
 
 		return { success: true };
